Move Air activities and traits to Activities\Air and Traits\Air, respectively
