+++ OpenRA.Mods.Cnc/Traits/Buildings/ProductionAirdrop.cs
@@ -11,7 +11,7 @@
 using System.Linq;
 using OpenRA.Mods.RA;
 using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Air;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
