+++ OpenRA.Mods.RA/AI/States/StateBase.cs
@@ -12,7 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Air;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Mods.RA.Buildings;
 using OpenRA.Traits;
