+++ OpenRA.Mods.RA/Activities/Air/FallToEarth.cs
@@ -1,4 +1,4 @@
-﻿#region Copyright & License Information
+#region Copyright & License Information
 /*
  * Copyright 2007-2014 The OpenRA Developers (see AUTHORS)
  * This file is part of OpenRA, which is free software. It is made
@@ -10,31 +10,12 @@
 
 using System.Linq;
 using OpenRA.GameRules;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Air
+namespace OpenRA.Mods.RA.Activities
 {
-	class FallsToEarthInfo : ITraitInfo
-	{
-		[WeaponReference]
-		public readonly string Explosion = "UnitExplode";
-
-		public readonly bool Spins = true;
-		public readonly bool Moves = false;
-		public readonly WRange Velocity = new WRange(43);
-
-		public object Create(ActorInitializer init) { return new FallsToEarth(init.self, this); }
-	}
-
-	class FallsToEarth
-	{
-		public FallsToEarth(Actor self, FallsToEarthInfo info)
-		{
-			self.QueueActivity(false, new FallToEarth(self, info));
-		}
-	}
-
-	class FallToEarth : Activity
+	public class FallToEarth : Activity
 	{
 		int acceleration = 0;
 		int spin = 0;
