+++ OpenRA.Mods.RA/Activities/Air/Fly.cs
@@ -9,9 +9,10 @@
 #endregion
 
 using System.Collections.Generic;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Air
+namespace OpenRA.Mods.RA.Activities
 {
 	public class Fly : Activity
 	{
