+++ OpenRA.Mods.RA/Activities/Air/HeliFly.cs
@@ -9,9 +9,10 @@
 #endregion
 
 using System.Collections.Generic;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Air
+namespace OpenRA.Mods.RA.Activities
 {
 	class HeliFly : Activity
 	{
