+++ OpenRA.Mods.RA/Activities/Air/HeliReturn.cs
@@ -9,10 +9,10 @@
 #endregion
 
 using System.Linq;
-using OpenRA.Mods.RA.Activities;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Air
+namespace OpenRA.Mods.RA.Activities
 {
 	public class HeliReturn : Activity
 	{
