+++ OpenRA.Mods.RA/Activities/Air/Land.cs
@@ -8,9 +8,10 @@
  */
 #endregion
 
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Air
+namespace OpenRA.Mods.RA.Activities
 {
 	public class Land : Activity
 	{
