--- OpenRA.Mods.RA/Air/FlyAwayOnIdle.cs
@@ -1,27 +1,0 @@
-﻿#region Copyright & License Information
-/*
- * Copyright 2007-2014 The OpenRA Developers (see AUTHORS)
- * This file is part of OpenRA, which is free software. It is made
- * available to you under the terms of the GNU General Public License
- * as published by the Free Software Foundation. For more information,
- * see COPYING.
- */
-#endregion
-
-using OpenRA.Mods.RA.Activities;
-using OpenRA.Traits;
-
-namespace OpenRA.Mods.RA.Air
-{
-	[Desc("Leave the map when idle.")]
-	class FlyAwayOnIdleInfo : TraitInfo<FlyAwayOnIdle> { }
-
-	class FlyAwayOnIdle : INotifyIdle
-	{
-		public void TickIdle(Actor self)
-		{
-			self.QueueActivity(new FlyOffMap());
-			self.QueueActivity(new RemoveSelf());
-		}
-	}
-}
