+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -80,7 +80,20 @@
   <ItemGroup>
     <Compile Include="Activities\CaptureActor.cs" />
     <Compile Include="Activities\Hunt.cs" />
-    <Compile Include="Air\TakeOff.cs" />
+    <Compile Include="Activities\Air\FallToEarth.cs" />
+    <Compile Include="Activities\Air\Fly.cs" />
+    <Compile Include="Activities\Air\FlyAttack.cs" />
+    <Compile Include="Activities\Air\FlyCircle.cs" />
+    <Compile Include="Activities\Air\FlyFollow.cs" />
+    <Compile Include="Activities\Air\FlyTimed.cs" />
+    <Compile Include="Activities\Air\HeliAttack.cs" />
+    <Compile Include="Activities\Air\HeliFly.cs" />
+    <Compile Include="Activities\Air\HeliLand.cs" />
+    <Compile Include="Activities\Air\HeliReturn.cs" />
+    <Compile Include="Activities\Air\Land.cs" />
+    <Compile Include="Activities\Air\ResupplyAircraft.cs" />
+    <Compile Include="Activities\Air\ReturnToBase.cs" />
+    <Compile Include="Activities\Air\TakeOff.cs" />
     <Compile Include="AI\AttackOrFleeFuzzy.cs" />
     <Compile Include="AI\BaseBuilder.cs" />
     <Compile Include="AI\HackyAI.cs" />
@@ -112,9 +125,6 @@
     <Compile Include="Activities\UnloadCargo.cs" />
     <Compile Include="Activities\Wait.cs" />
     <Compile Include="ActorExts.cs" />
-    <Compile Include="Air\Aircraft.cs" />
-    <Compile Include="Air\AttackHeli.cs" />
-    <Compile Include="Air\AttackPlane.cs" />
     <Compile Include="AI\SupportPowerDecision.cs" />
     <Compile Include="Crates\DuplicateUnitCrateAction.cs" />
     <Compile Include="Effects\GpsSatellite.cs" />
@@ -126,21 +136,6 @@
     <Compile Include="Infiltration\InfiltrateForPowerOutage.cs" />
     <Compile Include="Player\PlaceBeacon.cs" />
     <Compile Include="EjectOnDeath.cs" />
-    <Compile Include="Air\FallsToEarth.cs" />
-    <Compile Include="Air\Fly.cs" />
-    <Compile Include="Air\FlyAttack.cs" />
-    <Compile Include="Air\FlyCircle.cs" />
-    <Compile Include="Air\FlyTimed.cs" />
-    <Compile Include="Air\HeliAttack.cs" />
-    <Compile Include="Air\HeliFly.cs" />
-    <Compile Include="Air\HeliLand.cs" />
-    <Compile Include="Air\HeliReturn.cs" />
-    <Compile Include="Air\Helicopter.cs" />
-    <Compile Include="Air\Land.cs" />
-    <Compile Include="Air\Plane.cs" />
-    <Compile Include="Air\ReturnOnIdle.cs" />
-    <Compile Include="Air\ReturnToBase.cs" />
-    <Compile Include="Air\TargetableAircraft.cs" />
     <Compile Include="AI\RushFuzzy.cs" />
     <Compile Include="AI\StateMachine.cs" />
     <Compile Include="AppearsOnRadar.cs" />
@@ -306,6 +301,15 @@
     <Compile Include="TargetableUnit.cs" />
     <Compile Include="ThrowsParticle.cs" />
     <Compile Include="TraitsInterfaces.cs" />
+    <Compile Include="Traits\Air\Aircraft.cs" />
+    <Compile Include="Traits\Air\AttackHeli.cs" />
+    <Compile Include="Traits\Air\AttackPlane.cs" />
+    <Compile Include="Traits\Air\FlyAwayOnIdle.cs" />
+    <Compile Include="Traits\Air\FallsToEarth.cs" />
+    <Compile Include="Traits\Air\Helicopter.cs" />
+    <Compile Include="Traits\Air\Plane.cs" />
+    <Compile Include="Traits\Air\ReturnOnIdle.cs" />
+    <Compile Include="Traits\Air\TargetableAircraft.cs" />
     <Compile Include="TransformOnCapture.cs" />
     <Compile Include="TransformOnPassenger.cs" />
     <Compile Include="Transforms.cs" />
@@ -381,7 +385,6 @@
     <Compile Include="World\PathfinderDebugOverlay.cs" />
     <Compile Include="AttackBomber.cs" />
     <Compile Include="Render\WithCrateBody.cs" />
-    <Compile Include="Activities\FlyFollow.cs" />
     <Compile Include="Widgets\Logic\GameTimerLogic.cs" />
     <Compile Include="Widgets\Logic\ReplayControlBarLogic.cs" />
     <Compile Include="World\BuildableTerrainLayer.cs" />
@@ -428,7 +431,6 @@
     <Compile Include="Render\WithActiveAnimation.cs" />
     <Compile Include="SupportPowers\SpawnActorPower.cs" />
     <Compile Include="Render\RenderSimple.cs" />
-    <Compile Include="Air\FlyAwayOnIdle.cs" />
     <Compile Include="Buildings\ClonesProducedUnits.cs" />
     <Compile Include="Cloneable.cs" />
     <Compile Include="Widgets\ProductionPaletteWidget.cs" />
