+++ OpenRA.Mods.RA/Scripting/Properties/HelicopterProperties.cs
@@ -8,7 +8,8 @@
  */
 #endregion
 
-using OpenRA.Mods.RA.Air;
+using OpenRA.Mods.RA.Activities;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
