+++ OpenRA.Mods.RA/Scripting/Properties/TransportProperties.cs
@@ -10,7 +10,7 @@
 
 using System.Linq;
 using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Air;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
