+++ OpenRA.Mods.RA/Traits/Air/Helicopter.cs
@@ -12,10 +12,11 @@ using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
+using OpenRA.Mods.RA;
 using OpenRA.Mods.RA.Activities;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Air
+namespace OpenRA.Mods.RA.Traits
 {
 	class HelicopterInfo : AircraftInfo, IMoveInfo
 	{
